+++ CHANGELOG.md
@@ -112,6 +112,7 @@
      `Observable...` with `ObservableValue...`.
 * 💥 **\[core\]** Remove the unused class `arez.component.AbstractEntityReference`.
 * 💥 **\[core\]** Add tests covering the spy events `arez.spy.*Event`.
+* 💥 **\[core\]** Change the `arez.spy.Component*Event` spy events to use `ComponentInfo` rather than `Component`.
 
 ### [v0.104](https://github.com/arez/arez/tree/v0.104) (2018-08-01)
 [Full Changelog](https://github.com/arez/arez/compare/v0.103...v0.104)